Make polyply installable with pip
To do so I had to change a bit the organization of the repository. I
moved the python module in the new "polyply" directory, but I let the
metadata files (README, etc) at the top level of the repository for
github to find them.

I moved polyply.py in the new "bin" directory. I renamed the script
"polyply" instead of "polyply.py" as it was referred at under the former
name in the README, and because it was causing conflicts with the module
being called polyply.

Initially by mistake, but also for aesthetic reasons, I renamed the
module polyply. Importing the module is therefore done with

    import polyply

instead of

    import Martini_PolyPly

With this commit, PolyPly can be installed using pip by calling

    pip install <directory where setup.py is>

More easily, when in the Martini_PolyPly directory, the module can be
installed with

    pip install ./

The development version can be installed directly from github with

    pip install git+https://github.com/fgrunewald/Martini_PolyPly.git#polyply

The polyply module then becomes available for import with

    import polyply

The polyply script is made accessible in the research PATH, and can be
called from anywhere.