Fix the path to the data
The path to the distributed data was making wrong assumptions regarding
the location of the polyply script relative to the data. This commit
creates a `polyply.DATA_PATH` variable that should always point toward
the right place. Update the variable is the data are moved.

Also, this commit replace path concatenation (using + on strings) by
`os.path.join`. The latter method makes it more robust to join path. It,
for instance, avoids to make assumption on the presence or not of a / at
the end of the path elements.