Adapt the tests to run on a different environment
This commit adapt test.sh to called polyply with the installed version.
It also remove the direct calls to "python3.5" in the same file.

Calling a specific version of python can make things break when the user
uses a different version of python, or when using virtual/conda
environment.

Also, calling "python polyply" will break as it circumvent the research
PATH, and therefore does not know where to look for the script.